requirements.txt
matplotlib
paho-mqtt
PyQt5
icecream 